Extract version serializer factory to allow customizing the behavior
